gf_backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0
pytz==2022.2.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2022.2
